@@ -1,86 +1,72 @@
 {
-  "name": "customreactboilerplate",
-  "version": "1.0.2",
+  "name": "custom-front-end",
+  "version": "1.0.0",
   "description": "",
   "main": "index.js",
   "scripts": {
-    "clean": "rimraf dist",
-    "start": "webpack-dev-server --mode development",
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "format": "prettier --write \"src/**/*.js\"",
-    "eslint-fix": "eslint --fix \"src/**/*.js\"",
-    "build": "npm run clean && webpack  --mode=production --progress --profile --colors --display-error-details"
+    "start": "webpack serve  --mode development",
+    "build:dev": "rimraf dist && webpack --mode=development",
+    "build:prod": "rimraf dist && webpack --progress --mode=production --node-env=production",
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "keywords": [
-    "react",
-    "reactjs",
-    "babel6",
-    "boilerplate",
-    "redux",
-    "hot",
-    "reload",
-    "hmr",
-    "live",
-    "edit",
-    "webpack"
-  ],
-  "author": "Himanshu Sharma <hmnshu26@gmail.com> (http://hmnshu.co)",
+  "keywords": [],
+  "author": "",
   "license": "ISC",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/hmnshushrma/react-custom-boilerplate.git"
-  },
-  "bugs": "https://github.com/hmnshushrma/react-custom-boilerplate/issues",
   "dependencies": {
-    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
-    "@babel/runtime": "^7.4.5",
-    "axios": "^0.19.0",
-    "babel-polyfill": "^6.26.0",
-    "clean-webpack-plugin": "^3.0.0",
-    "mini-css-extract-plugin": "^0.8.0",
-    "optimize-css-assets-webpack-plugin": "^5.0.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.1.0",
-    "redux": "^4.0.1",
-    "redux-devtools-extension": "^2.13.9",
-    "redux-logger": "^3.0.6",
-    "redux-thunk": "^2.3.0",
-    "uuid": "^8.3.2",
-    "webpack-bundle-analyzer": "^3.4.1"
+    "@babel/plugin-proposal-class-properties": "^7.18.6",
+    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
+    "@babel/plugin-transform-runtime": "^7.19.1",
+    "@reduxjs/toolkit": "^1.8.5",
+    "antd": "^4.23.4",
+    "axios": "^0.27.2",
+    "jwt-decode": "^3.1.2",
+    "node-sass": "^9.0.0",
+    "react": "^18.2.0",
+    "react-apexcharts": "^1.4.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.0.5",
+    "redux": "^4.2.0",
+    "redux-persist": "^6.0.0",
+    "redux-saga": "^1.2.1",
+    "regenerator-runtime": "^0.13.9",
+    "save-dev": "^0.0.1-security",
+    "styled-components": "^5.3.6",
+    "uuid": "^9.0.0",
+    "uuidv4": "^6.2.13"
   },
   "devDependencies": {
-    "@babel/core": "^7.4.5",
-    "@babel/plugin-proposal-class-properties": "^7.5.5",
-    "@babel/plugin-transform-runtime": "^7.4.4",
-    "@babel/preset-env": "^7.4.5",
-    "@babel/preset-react": "^7.0.0",
-    "@material-ui/core": "^4.1.2",
-    "@material-ui/icons": "^4.2.1",
-    "babel-eslint": "^10.0.1",
-    "babel-loader": "^8.0.6",
-    "css-loader": "^3.0.0",
-    "dotenv": "^8.0.0",
-    "eslint": "^5.16.0",
-    "eslint-config-prettier": "^6.1.0",
-    "eslint-config-react": "^1.1.7",
-    "eslint-config-standard": "^12.0.0",
-    "eslint-config-standard-jsx": "^6.0.2",
-    "eslint-loader": "^2.1.2",
-    "eslint-plugin-import": "^2.17.3",
-    "eslint-plugin-node": "^9.1.0",
-    "eslint-plugin-promise": "^4.1.1",
-    "eslint-plugin-react": "^7.13.0",
-    "eslint-plugin-standard": "^4.0.0",
-    "html-webpack-plugin": "^3.2.0",
-    "node-sass": "^4.12.0",
-    "prettier": "1.18.2",
-    "react-router-dom": "^5.0.1",
-    "sass-loader": "^7.1.0",
-    "style-loader": "^0.23.1",
-    "url-loader": "^2.0.1",
-    "webpack": "^4.33.0",
-    "webpack-cli": "^3.3.4",
-    "webpack-dev-server": "^3.7.1"
+    "@babel/core": "^7.14.6",
+    "@babel/eslint-parser": "^7.19.1",
+    "@babel/preset-env": "^7.14.5",
+    "@babel/preset-react": "^7.14.5",
+    "@svgr/webpack": "^6.4.0",
+    "babel-eslint": "^10.1.0",
+    "babel-loader": "^8.2.2",
+    "compression-webpack-plugin": "^10.0.0",
+    "css-loader": "^5.2.6",
+    "css-minimizer-webpack-plugin": "^5.0.1",
+    "dotenv": "^10.0.0",
+    "eslint": "^7.32.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-config-standard": "^16.0.3",
+    "eslint-config-standard-jsx": "^10.0.0",
+    "eslint-config-standard-react": "^11.0.1",
+    "eslint-plugin-import": "^2.23.4",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-prettier": "^3.4.0",
+    "eslint-plugin-promise": "^5.1.0",
+    "eslint-plugin-react": "^7.24.0",
+    "eslint-webpack-plugin": "^3.2.0",
+    "file-loader": "^6.2.0",
+    "html-webpack-plugin": "^5.3.1",
+    "mini-css-extract-plugin": "^1.6.0",
+    "prettier": "2.7.1",
+    "react-router-dom": "^6.4.1",
+    "sass-loader": "^12.1.0",
+    "standard": "^16.0.3",
+    "style-loader": "^2.0.0",
+    "webpack": "^5.74.0",
+    "webpack-cli": "^4.10.0",
+    "webpack-dev-server": "^4.15.1"
   }
 }
