@@ -1,132 +1,176 @@
-const webpack = require('webpack')
-const dotenv = require('dotenv')
 const path = require('path')
-const HtmlWebPackPlugin = require('html-webpack-plugin')
-const rootDir = path.resolve(__dirname, '.')
-const srcDir = path.resolve(__dirname, '.', 'src')
-const distDir = path.resolve(__dirname, '.', 'dist')
+const webpack = require('webpack')
+const HtmlWebpackPlugin = require('html-webpack-plugin')
 const MiniCssExtractPlugin = require('mini-css-extract-plugin')
-const devMode = process.env.NODE_ENV !== 'production'
-const OptimizeCSSAssetsPlugin = require('optimize-css-assets-webpack-plugin')
-const TerserJSPlugin = require('terser-webpack-plugin')
+const ESLintPlugin = require('eslint-webpack-plugin')
+const TerserPlugin = require('terser-webpack-plugin')
+const CompressionPlugin = require('compression-webpack-plugin')
+const CssMinimizerPlugin = require('css-minimizer-webpack-plugin')
+//
+const isProduction = process.env.NODE_ENV === 'production'
+const dotenv = require('dotenv')
+const stylesHandler = isProduction
+  ? MiniCssExtractPlugin.loader
+  : 'style-loader'
 
-module.exports = () => {
-  const env = dotenv.config().parsed
-  const envKeys = Object.keys(env).reduce((prev, next) => {
-    prev[`process.env.${next}`] = JSON.stringify(env[next])
-    return prev
-  }, {})
-  return {
-    context: rootDir,
-    entry: './src/index.js',
-    output: {
-      path: distDir,
-      publicPath: '/',
-      filename: '[hash].js'
-    },
-    devServer: {
-      contentBase: rootDir,
-      publicPath: '/',
-      historyApiFallback: true,
-      hot: true,
-      open: true
-      // needed for phone testing
-      // host:'0.0.0.0',
-      // port:'8080'
-    },
-    devtool: "#eval-source-map",
-    resolve: {
-      extensions: ['*', '.js', '.jsx'],
-      alias: {
-        Styles: path.resolve(__dirname, 'src/styles/'),
-        Assets: path.resolve(__dirname, 'src/assets/'),
-        Redux: path.resolve(__dirname, 'src/redux/'),
-        Views: path.resolve(__dirname, 'src/views/'),
-        Hoc: path.resolve(__dirname, 'src/hoc/'),
-        Routes: path.resolve(__dirname, 'src/routes/'),
-        FormConstants: path.resolve(__dirname, 'src/FormConstants/'),
-        Util: path.resolve(__dirname, 'src/util/')
-      }
-    },
-    module: {
-      rules: [
-        {
-          test: /\.(js|jsx)$/,
-          exclude: /node_modules/,
-          use: ['babel-loader', 'eslint-loader'], // include eslint-loader,
-          include: path.resolve(__dirname, './', 'src')
-        },
-        {
-          test: /\.(jpe?g|png|gif|woff|woff2|eot|ttf|otf|svg)(\?[a-z0-9=.]+)?$/,
-          loader: 'url-loader'
-        },
-        {
-          test: /\.s?[ac]ss$/,
-          use: [
-            MiniCssExtractPlugin.loader,
-            { loader: 'css-loader', options: { url: false, sourceMap: true } },
-            { loader: 'sass-loader', options: { sourceMap: true } }
-          ]
-        }
-      ]
-    },
-    optimization: {
-      minimize: true,
-      minimizer: [
-        new TerserJSPlugin({}),
-        new OptimizeCSSAssetsPlugin({
-          cssProcessor: require('cssnano'),
-          cssProcessorOptions: {
-            discardComments: {
-              removeAll: true
-            },
-            safe: true
+const env = dotenv.config().parsed
+const envKeys = Object.keys(env).reduce((prev, next) => {
+  prev[`process.env.${next}`] = JSON.stringify(env[next])
+  return prev
+}, {})
+
+const config = {
+  entry: './src/index.js',
+  output: {
+    path: path.resolve(__dirname, 'dist'),
+    // filename: '[chunkhash].js',
+    publicPath: '/',
+    filename: '[name].js',
+    chunkFilename: '[name].bundle.js'
+  },
+  devServer: {
+    hot: true,
+    // host:'0.0.0.0',
+    port: 3000,
+    historyApiFallback: true,
+    static: {
+      directory: path.resolve(__dirname, '..', './dist')
+    }
+  },
+  devtool: 'eval-source-map',
+  optimization: {
+    mangleExports: true,
+    minimizer: [
+      new TerserPlugin({
+        terserOptions: {
+          output: {
+            comments: false
           },
-          canPrint: false
-        })
-      ],
-      splitChunks: {
-        minSize: 30000,
-        maxSize: 0,
-        minChunks: 1,
-        automaticNameDelimiter: '~',
-        name: true,
-        cacheGroups: {
-          default: false,
-          vendors: false,
-          vendors: {
-            chunks: 'all',
-            test: /[\\/]node_modules[\\/]/,
-            priority: -10
-          },
-          common: {
-            chunks: 'async',
-            minChunks: 2,
-            name: 'common',
-            priority: -20,
-            reuseExistingChunk: true
+          compress: {
+            // The following is required for dead-code the removal
+            // check in React DevTools
+            //
+            // default
+            unused: true,
+            dead_code: true,
+            conditionals: true,
+            evaluate: true
           }
+        },
+        extractComments: false
+      })
+    ],
+    splitChunks: {
+      cacheGroups: {
+        reactVendor: {
+          test: /[\\/]node_modules[\\/](react|react-dom|react-router-dom)[\\/]/,
+          name: 'vendor-react',
+          chunks: 'all'
         }
       }
-    },
-    plugins: [
-      new HtmlWebPackPlugin({
-        // where to find the html template
-        template: path.resolve(__dirname, 'index.html'),
-        // where to put the generated file
-        path: distDir,
-        // the output file name
-        filename: 'index.html',
-        files: {
-          css: ['styles.css'],
-          js: ['bundle.js']
-        }
-      }),
-      new MiniCssExtractPlugin({
-        filename: '[hash].css'
-      }),
-      new webpack.DefinePlugin(envKeys)
-    ],
-    mode: devMode ? 'development' : 'production'
+    }
+  },
+  plugins: [
+    new webpack.HotModuleReplacementPlugin(),
+    new HtmlWebpackPlugin({
+      template: path.resolve(__dirname, 'index.html'),
+      filename: 'index.html'
+    }),
+    new MiniCssExtractPlugin({
+      filename: '[chunkhash].css'
+    }),
+    new webpack.DefinePlugin(envKeys),
+    new webpack.LoaderOptionsPlugin({
+      minimize: true,
+      debug: false,
+      options: {
+        context: __dirname
+      }
+    }),
+    new ESLintPlugin({
+      extensions: ['js', 'jsx'],
+      exclude: ['/node_modules/']
+    }),
+    new CompressionPlugin({
+      test: /\.js(\?.*)?$/i
+    })
+  ],
+  cache: {
+    type: 'filesystem',
+    allowCollectingMemory: true
+  },
+  module: {
+    rules: [
+      {
+        test: /\.(js|jsx)$/,
+        exclude: /node_modules/,
+        use: ['babel-loader']
+      },
+      {
+        test: /\.css$/i,
+        use: [MiniCssExtractPlugin.loader, stylesHandler, 'css-loader']
+      },
+      {
+        test: /\.s[ac]ss$/i,
+        use: [
+          MiniCssExtractPlugin.loader,
+          stylesHandler,
+          'css-loader',
+          'sass-loader'
+        ]
+      },
+      {
+        test: /\.(eot|svg|ttf|woff|woff2|png|jpg|gif)$/i,
+        type: 'asset'
+      }
+    ]
   }
 }
+module.exports = () => {
+  if (isProduction) {
+    config.mode = 'production'
+    config.plugins = prodPlugins
+    config.optimization = chunks
+  } else {
+    config.mode = 'development'
+  }
+  return config
+}
+
+const chunks = {
+  minimizer: ['...', new CssMinimizerPlugin()],
+  splitChunks: {
+    cacheGroups: {
+      reactVendor: {
+        test: /[\\/]node_modules[\\/](react|react-dom|react-router-dom)[\\/]/,
+        name: 'vendor-react',
+        chunks: 'all'
+      }
+    }
+  }
+}
+const prodPlugins = [
+  new MiniCssExtractPlugin(),
+  // new HtmlWebpackPlugin({
+  //   template: path.resolve(__dirname, 'index.html'),
+  //   filename: 'index.html'
+  // }),
+  new MiniCssExtractPlugin({
+    filename: '[chunkhash].css'
+  }),
+  new webpack.DefinePlugin(envKeys),
+  new webpack.LoaderOptionsPlugin({
+    minimize: true,
+    debug: false,
+    options: {
+      context: __dirname
+    }
+  }),
+  new ESLintPlugin({
+    extensions: ['js', 'jsx'],
+    exclude: ['/node_modules/']
+  }),
+  new CompressionPlugin({
+    test: /\.js(\?.*)?$/i
+  })
+]
